requirements.txt
# Core Machine Learning
torch==2.8.0
torchvision==0.18.0 
torchinfo==1.8.0

# Configuration & Optimization
hydra-core==1.3.2
optuna==4.6.0
omegaconf==2.3.0

# Experiment Tracking & Viz
wandb==0.23.1
seaborn==0.13.2
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
pillow>=9.5.0
numpy>=1.24.0
imagededup==0.3.3.post2
imagehash==4.3.2